requirements.txt
absl-py==2.0.0
attrs==23.2.0
boto3==1.33.13
botocore==1.33.13
cachetools==5.3.2
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.3.2
clearml==1.13.2
click==8.1.7
colorama==0.4.6
cycler==0.10.0
docopt==0.6.2
filelock==3.12.2
fire==0.5.0
fonttools==4.38.0
fsspec==2023.1.0
furl==2.1.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==0.4.6
grpcio==1.60.0
huggingface-hub==0.16.4
idna==2.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
jmespath==1.0.1
jsonschema==4.17.3
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
orderedmultidict==1.0.1
packaging==23.2
pandas==1.3.5
pathlib2==2.3.7.post1
Pillow==9.5.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
psutil==5.9.7
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
PyJWT==2.8.0
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.6
rsa==4.9
s3transfer==0.8.2
sahi==0.11.15
scipy==1.7.3
seaborn==0.12.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
supervision==0.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==1.26.18
Werkzeug==2.2.3
yarg==0.1.9
yolov5==7.0.13
yolov8==0.0.2
zipp==3.15.0
